pyproject.toml
[tool.poetry]
name = "oop_course_project"
version = "0.1.0"
description = ""
authors = ["Dizzzmas"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask_sqlalchemy = "^2.4.1"
JetKit-Flask = "^9.0.1"
flask_migrate = "^2.5.3"
nplusone = "^1.0.0"
psycopg2-binary = "^2.8.5"
flask_script = "^2.0.6"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
faker = "^4.0.3"
factory-boy = "^2.12.0"
pytest = "^5.4.1"
pytest_factoryboy = "^2.0.3"
pytest_postgresql = "^2.3.0"
pytest-flask-sqlalchemy = "^1.0.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
